Trace timeline: toggle success/failure paths, move summary above filters
- Replace tri-state path filter (all/success/failure) with independent
  success and failure toggle buttons
- Add colored icons to the toggle labels (green checkmark, red X)
- Move the summary bar above the filter buttons
- Move fixture module from src/ to test/
