+++ ts/packages/grammarTools/ui/dev/index.html
@@ -117,7 +117,7 @@
     </section>
 
     <script type="module">
-      import { FixtureBackend } from "../src/index.ts";
+      import { FixtureBackend } from "../test/fixture/fixtureBackend.ts";
       // Import to register all custom elements
       import "../src/index.ts";
