+++ ts/packages/grammarTools/ui/test/components/gt-rule-list.test.ts
@@ -3,7 +3,7 @@
 
 import { fixture, html, expect } from "@open-wc/testing";
 import type { GtRuleList } from "../../src/gt-rule-list.js";
-import { FixtureBackend } from "../../src/fixture/fixtureBackend.js";
+import { FixtureBackend } from "../fixture/fixtureBackend.js";
 
 // Import to register element
 import "../../src/gt-rule-list.js";
